Hlo module splitting
This change introduces the ability to split a single HLO module into multiple modules, for the separate compilation effort.

Splitting strategy assigns each computation in the call graph (starting from the entry computation) to its own separate module. Calls between computations that end up in different modules are replaced by stubs, which are implemented as custom-calls. An HloLinkingManifest is generated alongside the split modules, containing information to link the modules back together by replacing stubs with calls to the actual computations in other modules.

This change also adds helper methods to assign deterministic instruction and computation IDs, which are used to generate stable module names, and to replace called computations when inserting stubs.

PiperOrigin-RevId: 813708610