[XLA] Implement a `TiledHloSchedule` that transposes the iteration order over the non-contracting dimensions of a `dot`.
A concrete use case when such a schedule is useful is when we have a matrix
multiplication such that a chunk of shape `(block_m, k)` of the left-hand
side argument fully fits into L2. The transposed iteration order will step
through the `n` dimension first, allowing to hit L2 cache more often when
loading tiles of the left-hand side.

This schedule is intentionally restricted at the moment in order to unblock
launching the generic Triton emitter for GEMMs.

PiperOrigin-RevId: 816692263