Set call result shardings to the out shardings of func that is created or found from cache.
It is a no op for `dedupFunctionsFully` is false which is also the default.

outShardings is the the output shardings of the named computation at hand. However, if dedupFunctionsFully true, the func we pick from `createFuncOpOrGetFromCache`, which is the func the call will actually be calling, may have a different output sharding than the named computation, and call result sharding should be set to the output sharding it calls. For example,

namedComputation1(foo): insharding={"x"} outsharding={"y"}
namedComputation2(foo): insharding={"x"} outsharding={"z"}

call1 to namedComputation1
call2 to namedComputation2

When dedupFunctionsFully is false, we have separate instances of foo as their outshardings are different.

func foo1 insharding={"x"} outsharding={"y"} {...}
func foo2 insharding={"x"} outsharding={"z"} {...}
call1 to foo_1 resultsharding={"y"}
call2 to foo_2 resultsharding={"z"}

When dedupFunctionsFully is true, we do not have separate instance of foo, we need to pick either namedComputation1 or namedComputation2, say we pick namedComputation1, hence it becomes:

func foo insharding={"x"} outsharding={"y"} {...}
call1 to foo resultsharding={"y"}
call2 to foo resultsharding={"y"}

As a result, call2 should have a resultsharding={"y"} since it is calling foo, instead of the out sharding of namedComputation2 which is {"z"}.

PiperOrigin-RevId: 818642021