Use ABSL_LOG instead of LOG to avoid name collisions with other projects.
It's easy for the LOG name to collide with other projects, which then
causes an error in the TSL logging because it no longer refers to the
Absl symbol.

To prevent this, use the `ABSL_` prefixed name. This allows a downstream
project to define `LOG`, while letting TSL refer to the symbol it needs.

PiperOrigin-RevId: 826212189