Enable new Triton emitter features and switch to nested GEMM fusion in tests.
The generic emitter uses this input and is the new default path. The legacy emitter will be removed soon.

Change fusion kind from `__triton_scaled_dot_fusion` to `__triton_nested_gemm_fusion` in one test so that `GemmFusionAutotunerRewriterVisitor::HandleFusion` does not add a `triton_gemm_config` which would then trigger NestGemmFusion to double nest and break tiling analysis. The plan here is to remove `__triton_scaled_dot_fusion` in the near future, so we don't need to do anything else.

PiperOrigin-RevId: 828967396